refactor(recommended-questions): move findOrInsert and generation into service
