+++ src/recommended-questions/recommended-questions.service.ts
@@ -1,4 +1,5 @@
 import { Injectable } from '@nestjs/common';
+import { InjectRepository } from '@nestjs/typeorm';
 import { ConversationsService } from 'omniboxd/conversations/conversations.service';
 import { NamespaceResourcesService } from 'omniboxd/namespace-resources/namespace-resources.service';
 import {
@@ -7,9 +8,11 @@ import {
   RecommendQuestionsResponseDto,
   RecommendResourceDto,
 } from 'omniboxd/recommended-questions/dto/recommend-questions.dto';
+import { RecommendedQuestion } from 'omniboxd/recommended-questions/entities/recommended-question.entity';
 import { Resource } from 'omniboxd/resources/entities/resource.entity';
 import { TagService } from 'omniboxd/tag/tag.service';
 import { WizardAPIService } from 'omniboxd/wizard-api/wizard-api.service';
+import { QueryFailedError, Repository } from 'typeorm';
 
 const RECENT_RESOURCES_COUNT = 5;
 const RECENT_TAGS_COUNT = 10;
@@ -17,16 +20,55 @@ const RECENT_QUESTIONS_COUNT = 5;
 const DEFAULT_MAX_QUESTIONS = 3;
 const RESOURCE_CONTENT_MAX_LENGTH = 500;
 
+const PG_UNIQUE_VIOLATION = '23505';
+
+function isUniqueViolation(err: unknown): boolean {
+  return (
+    err instanceof QueryFailedError &&
+    (err.driverError as { code?: string } | undefined)?.code ===
+      PG_UNIQUE_VIOLATION
+  );
+}
+
 @Injectable()
 export class RecommendedQuestionsService {
   constructor(
+    @InjectRepository(RecommendedQuestion)
+    private readonly recommendedQuestionsRepository: Repository<RecommendedQuestion>,
     private readonly namespaceResourcesService: NamespaceResourcesService,
     private readonly tagService: TagService,
     private readonly conversationsService: ConversationsService,
     private readonly wizardApiService: WizardAPIService,
   ) {}
 
-  async getRecommendedQuestions(
+  async findOrInsert(
+    namespaceId: string,
+    userId: string,
+    now: Date,
+  ): Promise<RecommendedQuestion> {
+    const record = await this.recommendedQuestionsRepository.findOne({
+      where: { namespaceId, userId },
+    });
+    if (record) {
+      return record;
+    }
+    try {
+      await this.recommendedQuestionsRepository.insert({
+        namespaceId,
+        userId,
+        scannedAt: now,
+      });
+    } catch (err) {
+      if (!isUniqueViolation(err)) {
+        throw err;
+      }
+    }
+    return await this.recommendedQuestionsRepository.findOneOrFail({
+      where: { namespaceId, userId },
+    });
+  }
+
+  async generateRecommendedQuestions(
     namespaceId: string,
     userId: string,
     maxQuestions: number = DEFAULT_MAX_QUESTIONS,
